test: add domain + application service tests
- all services have appropriate tests.
- update test helper functions

+++ src/test/java/antigravity/domain/service/DiscountCalculationServiceTests.java
@@ -1,7 +1,6 @@
 package antigravity.domain.service;
 
 import antigravity.domain.entity.Product;
-import antigravity.domain.service.DiscountCalculationService;
 import antigravity.exception.EntityNotFoundException;
 import antigravity.testutils.TestHelper;
 import org.junit.jupiter.api.Test;
@@ -9,7 +8,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 import static org.junit.jupiter.api.Assertions.*;
@@ -26,10 +24,10 @@ public class DiscountCalculationServiceTests {
         Product product = TestHelper.buildSampleProduct();
         List<Integer> promotionIds = Arrays.asList(1, 2);
 
-        BigDecimal discount = service.calculateDiscountAmount(product.getPrice(), promotionIds);
+        int discount = service.calculateDiscountAmount(product.getPrice(), promotionIds);
 
         // Applying existing promotions to product 1 = 62250
-        assertEquals(discount.intValue(), 62250.00);
+        assertEquals(discount, 62250);
     }
 
     @Test
@@ -38,11 +36,11 @@ public class DiscountCalculationServiceTests {
         List<Integer> firstPromotionId = Arrays.asList(1);
         List<Integer> secondPromotionId = Arrays.asList(2);
 
-        BigDecimal firstDiscount = service.calculateDiscountAmount(product.getPrice(), firstPromotionId);
-        BigDecimal secondDiscount = service.calculateDiscountAmount(product.getPrice(), secondPromotionId);
+        int firstDiscount = service.calculateDiscountAmount(product.getPrice(), firstPromotionId);
+        int secondDiscount = service.calculateDiscountAmount(product.getPrice(), secondPromotionId);
 
-        assertEquals(firstDiscount.intValue(), 30000.00);
-        assertEquals(secondDiscount.intValue(), 32250.00);
+        assertEquals(firstDiscount, 30000);
+        assertEquals(secondDiscount, 32250);
     }
 
     @Test
